コントローラをモデルごとに分岐
